refactor: secure 패키지를 auth 아래로 이동 후 encryption 으로 이름 변경
