+++ backend/src/main/java/codezap/auth/encryption/PasswordEncryptor.java
@@ -1,4 +1,4 @@
-package codezap.secure;
+package codezap.auth.encryption;
 
 public interface PasswordEncryptor {
     String encrypt(String plainPassword, String salt);
