+++ backend/src/main/java/codezap/auth/encryption/SaltGenerator.java
@@ -1,4 +1,4 @@
-package codezap.secure;
+package codezap.auth.encryption;
 
 @FunctionalInterface
 public interface SaltGenerator {
