+++ backend/src/test/java/codezap/auth/encryption/SHA2PasswordEncryptorTest.java
@@ -1,7 +1,6 @@
-package codezap.secure;
+package codezap.auth.encryption;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
