+++ backend/src/test/java/codezap/template/controller/TemplateControllerTest.java
@@ -40,10 +40,10 @@ import codezap.member.dto.MemberDto;
 import codezap.member.repository.FakeMemberRepository;
 import codezap.member.repository.MemberRepository;
 import codezap.member.service.MemberService;
-import codezap.secure.PasswordEncryptor;
-import codezap.secure.RandomSaltGenerator;
-import codezap.secure.SHA2PasswordEncryptor;
-import codezap.secure.SaltGenerator;
+import codezap.auth.encryption.PasswordEncryptor;
+import codezap.auth.encryption.RandomSaltGenerator;
+import codezap.auth.encryption.SHA2PasswordEncryptor;
+import codezap.auth.encryption.SaltGenerator;
 import codezap.tag.service.TemplateTagService;
 import codezap.template.dto.request.CreateSourceCodeRequest;
 import codezap.template.dto.request.CreateTemplateRequest;
